+++ JotunnLib/Managers/GUIManager.cs
@@ -3,12 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using Jotunn.Configs;
 using UnityEngine;
-using UnityEngine.EventSystems;
+using UnityEngine.Rendering;
 using UnityEngine.SceneManagement;
 using UnityEngine.UI;
 using Button = UnityEngine.UI.Button;
 using Image = UnityEngine.UI.Image;
-using Object = UnityEngine.Object;
 using Toggle = UnityEngine.UI.Toggle;
 
 namespace Jotunn.Managers
@@ -17,7 +16,7 @@ namespace Jotunn.Managers
     ///     Manager for handling anything GUI related. Provides Valheim style 
     ///     GUI elements as well as an anchor for custom GUI prefabs.
     /// </summary>
-    public class GUIManager : IManager, IPointerClickHandler
+    public class GUIManager : IManager
     {
         private static GUIManager _instance;
         /// <summary>
@@ -125,14 +124,12 @@ namespace Jotunn.Managers
         private bool GUIInStart = false;
 
         /// <summary>
-        ///     Event receiver for pointer click events
+        ///     Detect headless mode (aka dedicated server)
         /// </summary>
-        /// <param name="eventData"></param>
-        public void OnPointerClick(PointerEventData eventData)
+        /// <returns></returns>
+        public static bool IsHeadless()
         {
-#if DEBUG
-            Logger.LogMessage(eventData.GetObjectString());
-#endif
+            return SystemInfo.graphicsDeviceType == GraphicsDeviceType.Null;
         }
 
         /// <summary>
@@ -140,29 +137,38 @@ namespace Jotunn.Managers
         /// </summary>
         public void Init()
         {
-            GUIContainer = new GameObject("GUI");
-            GUIContainer.transform.SetParent(Main.RootObject.transform);
-            GUIContainer.layer = UILayer; // UI
-            var canvas = GUIContainer.AddComponent<Canvas>();
-            canvas.renderMode = RenderMode.ScreenSpaceOverlay;
-            canvas.sortingOrder = 1;
-            GUIContainer.GetComponent<RectTransform>().localScale = new Vector3(1, 1, 1);
-            GUIContainer.AddComponent<GraphicRaycaster>();
-            GUIContainer.AddComponent<CanvasScaler>();
-            GUIContainer.GetComponent<RectTransform>().anchorMin = new Vector2(0.5f, 0.5f);
-            GUIContainer.GetComponent<RectTransform>().anchorMax = new Vector2(0.5f, 0.5f);
-            GUIContainer.GetComponent<RectTransform>().anchoredPosition = new Vector2(0, 0);
-            GUIContainer.GetComponent<Canvas>().planeDistance = 0.0f;
-            GUIContainer.AddComponent<GuiScaler>().UpdateScale();
+            // Dont init on a dedicated server
+            if (!IsHeadless())
+            {
+                GUIContainer = new GameObject("GUI");
+                GUIContainer.transform.SetParent(Main.RootObject.transform);
+                GUIContainer.layer = UILayer; // UI
+                var canvas = GUIContainer.AddComponent<Canvas>();
+                canvas.renderMode = RenderMode.ScreenSpaceOverlay;
+                canvas.sortingOrder = 1;
+                GUIContainer.GetComponent<RectTransform>().localScale = new Vector3(1, 1, 1);
+                GUIContainer.AddComponent<GraphicRaycaster>();
+                GUIContainer.AddComponent<CanvasScaler>();
+                GUIContainer.GetComponent<RectTransform>().anchorMin = new Vector2(0.5f, 0.5f);
+                GUIContainer.GetComponent<RectTransform>().anchorMax = new Vector2(0.5f, 0.5f);
+                GUIContainer.GetComponent<RectTransform>().anchoredPosition = new Vector2(0, 0);
+                GUIContainer.GetComponent<Canvas>().planeDistance = 0.0f;
+                GUIContainer.AddComponent<GuiScaler>().UpdateScale();
 
-            SceneManager.sceneLoaded += SceneManager_sceneLoaded;
-            On.KeyHints.UpdateHints += ShowCustomKeyHint;
+                SceneManager.sceneLoaded += SceneManager_sceneLoaded;
+                On.KeyHints.UpdateHints += ShowCustomKeyHint;
+            }
         }
 
         internal void OnGUI()
         {
-            // Load valheim GUI assets
+            // No resource loading on a headless server
+            if (IsHeadless())
+            {
+                return;
+            }
 
+            // Load valheim GUI assets
             if (needsLoad && SceneManager.GetActiveScene().name == "start" && SceneManager.GetActiveScene().isLoaded)
             {
                 try