+++ JotunnLib/Managers/ItemManager.cs
@@ -346,9 +346,8 @@ namespace Jotunn.Managers
                         {
                             itemDrop.m_itemData.m_dropPrefab = customItem.ItemPrefab;
                         }
-                        objectDB.m_items.Add(customItem.ItemPrefab);
 
-                        Logger.LogInfo($"Added item {customItem} | Token: {customItem.ItemDrop.TokenName()}");
+                        RegisterItemInObjectDB(customItem.ItemPrefab);
                     }
                     catch (Exception ex)
                     {
@@ -367,6 +366,49 @@ namespace Jotunn.Managers
 
                 objectDB.UpdateItemHashes();
             }
+        }
+
+        /// <summary>
+        ///     Register a single item in the current ObjectDB.
+        ///     Also adds the prefab to the <see cref="PrefabManager"/> and <see cref="ZNetScene"/> if necessary.<br />
+        ///     No mock references are fixed.
+        /// </summary>
+        /// <param name="prefab"><see cref="GameObject"/> with an <see cref="ItemDrop"/> component to add to the <see cref="ObjectDB"/></param>
+        public void RegisterItemInObjectDB(GameObject prefab)
+        {
+            var itemDrop = prefab.GetComponent<ItemDrop>();
+            if (itemDrop == null)
+            {
+                throw new Exception($"Prefab {prefab.name} has no ItemDrop component attached");
+            }
+
+            var objectDB = ObjectDB.instance;
+            if (objectDB == null)
+            {
+                throw new Exception($"ObjectDB is not instantiated");
+            }
+
+            var hash = prefab.name.GetStableHashCode();
+            if (objectDB.m_itemByHash.ContainsKey(hash))
+            {
+                Logger.LogInfo($"Already added item {prefab.name}");
+            }
+            else
+            {
+                if (!PrefabManager.Instance.Prefabs.ContainsKey(prefab.name))
+                {
+                    PrefabManager.Instance.AddPrefab(prefab);
+                }
+                if (ZNetScene.instance != null && !ZNetScene.instance.m_namedPrefabs.ContainsKey(hash))
+                {
+                    PrefabManager.Instance.RegisterToZNetScene(prefab);
+                }
+
+                objectDB.m_items.Add(prefab);
+                objectDB.m_itemByHash.Add(hash, prefab);
+            }
+
+            Logger.LogInfo($"Added item {prefab.name} | Token: {itemDrop.TokenName()}");
         }
 
         private void RegisterCustomRecipes(ObjectDB objectDB)
@@ -560,15 +602,15 @@ namespace Jotunn.Managers
 
             if (SceneManager.GetActiveScene().name == "start")
             {
-                var isValid = self.IsValid();
-                ItemDropMockFix.Switch(!isValid);
+                /*var isValid = self.IsValid();
+                ItemDropMockFix.Switch(!isValid);*/
 
-                if (isValid)
-                {
+                /*if (isValid)
+                {*/
                     RegisterCustomItems(self);
 
                     self.UpdateItemHashes();
-                }
+                //}
             }
         }
 
@@ -576,7 +618,8 @@ namespace Jotunn.Managers
         {
             orig(self, other);
 
-            if (SceneManager.GetActiveScene().name == "start" && self.IsValid())
+            //if (SceneManager.GetActiveScene().name == "start" && self.IsValid())
+            if (SceneManager.GetActiveScene().name == "start")
             {
                 OnItemsRegisteredFejd?.SafeInvoke();
             }
@@ -593,25 +636,26 @@ namespace Jotunn.Managers
 
             if (SceneManager.GetActiveScene().name == "main")
             {
-                var isValid = self.IsValid();
-                ItemDropMockFix.Switch(!isValid);
+                /*var isValid = self.IsValid();
+                ItemDropMockFix.Switch(!isValid);*/
 
-                if (isValid)
-                {
+                /*if (isValid)
+                {*/
                     RegisterCustomItems(self);
                     RegisterCustomRecipes(self);
                     RegisterCustomStatusEffects(self);
                     RegisterCustomItemConversions();
 
                     self.UpdateItemHashes();
-                }
+                //}
             }
         }
         private void InvokeOnItemsRegistered(On.ObjectDB.orig_Awake orig, ObjectDB self)
         {
             orig(self);
 
-            if (SceneManager.GetActiveScene().name == "main" && self.IsValid())
+            //if (SceneManager.GetActiveScene().name == "main" && self.IsValid())
+            if (SceneManager.GetActiveScene().name == "main")
             {
                 OnItemsRegistered?.SafeInvoke();
             }